SEQUOIADBMAINSTREAM-8293 【优化】【dms】【代码检视】部分场景下没有打印正确的错误码，影响诊断日志分析

+++ SequoiaDB/engine/dms/dmsStatSUMgr.cpp
@@ -740,14 +740,19 @@ namespace engine
          }
 
          tmpRC = _deleteCollectionStat( boMatcher, cb, NULL ) ;
-         PD_LOG( PDWARNING,
-                 "Failed to drop collection statistics when dropping "
-                 "collection space [%s], rc: %d", pCSName, tmpRC ) ;
-
+         if ( tmpRC )
+         {
+            PD_LOG( PDWARNING,
+                    "Failed to drop collection statistics when dropping "
+                    "collection space [%s], rc: %d", pCSName, tmpRC ) ;
+         }
          tmpRC = _deleteIndexStat( boMatcher, cb, NULL ) ;
-         PD_LOG( PDWARNING,
-                 "Failed to delete index statistics when dropping "
-                 "collection space [%s], rc: %d", pCSName, tmpRC ) ;
+         if ( tmpRC )
+         {
+            PD_LOG( PDWARNING,
+                    "Failed to delete index statistics when dropping "
+                    "collection space [%s], rc: %d", pCSName, tmpRC ) ;
+         }
       }
 
    done :
